finished refactoring voxel client
